backend/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
requests==2.31.0
flask-cors==3.0.10
google-generativeai==0.3.0
python-dotenv==0.19.0
eventlet==0.33.3
